[ALPS03763450] DISP: Fix coverity issue
[Detail]
Defect id: 4245487, 4245469, 4193873

MTK-Commit-Id: 12e98e00d3ba884ec63cdd0bbb3c6aab02413f22

CR-Id: ALPS03763450
Feature: Others

+++ drivers/misc/mediatek/video/mt6765/videox/debug.c
@@ -514,7 +514,7 @@ static int alloc_buffer_from_dma(size_t size, struct test_buf_info *buf_info)
 		goto out;
 	}
 	disp_ion_get_mva(ion_display_client, ion_display_handle,
-		(unsigned int *)&mva, DISP_M4U_PORT_DISP_WDMA0);
+		(unsigned long *)&mva, DISP_M4U_PORT_DISP_WDMA0);
 
 out:
 #endif
